+++ gui/user_variable_dialogs/blinker_gui.json
